# api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_iarsham_teacher-tool-api_internal_entities.UserRequest:
    properties:
      password:
        example: 1qaz2wsx
        maxLength: 32
        minLength: 8
        type: string
      phone:
        example: "+9891154326250"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - password
    - phone
    type: object
...
  response.BadRequest:
    properties:
      error:
        example: bad request
        type: string
    type: object
...
  response.InternalServerError:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  response.UserAlreadyExists:
    properties:
      error:
        example: user already exists
        type: string
    type: object
  response.UserCreated:
    properties:
      response:
        example: user created
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: arshamdev2001@gmail.com
    name: Arsham Roshannejad
    url: arsham.cloudarshamdev2001@gmail.com
  description: API for Teacher Tools application that provides various endpoints for
    managing data.
  license:
    name: MIT
    url: https://www.mit.edu/~amini/LICENSE.md
  termsOfService: http://swagger.io/terms/
  title: Teacher-Tools-API
  version: 0.0.0
paths:
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User data
        in: body
        name: userRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_iarsham_teacher-tool-api_internal_entities.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.UserCreated'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequest'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.UserAlreadyExists'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerError'
      summary: Register
      tags:
      - Auth
...
swagger: "2.0"